wip: create question and uint -> uint64

+++ entities/user.go
@@ -4,7 +4,7 @@ import "time"
 
 type User struct {
 	//USERID AUTO GENERATE
-	UserId  uint `gorm:"primaryKey; autoIncrement"`
+	UserId  uint64 `gorm:"primaryKey; autoIncrement"`
 	Account		string `gorm:"type:varchar(50); unique; notNull;"`
 	Pwd		string `gorm:"notNull"`
 	Last_login time.Time `gorm:"autoCreateTime"`
